Rename package from com.braintreepayments.browserswitch to com.braintreepayments.api.
