+++ browser-switch/src/androidTest/java/com/braintreepayments/api/test/TestBrowserSwitchFragment.java
@@ -1,0 +1,23 @@
+package com.braintreepayments.api.test;
+
+import android.net.Uri;
+
+import androidx.annotation.Nullable;
+
+import com.braintreepayments.api.BrowserSwitchResult;
+
+public class TestBrowserSwitchFragment extends BrowserSwitchFragment {
+
+    public boolean onBrowserSwitchResultCalled = false;
+    public int requestCode;
+    public BrowserSwitchResult result;
+    public Uri returnUri;
+
+    @Override
+    public void onBrowserSwitchResult(int requestCode, BrowserSwitchResult result, @Nullable Uri returnUri) {
+        onBrowserSwitchResultCalled = true;
+        this.requestCode = requestCode;
+        this.result = result;
+        this.returnUri = returnUri;
+    }
+}
