+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchException.java
@@ -1,0 +1,8 @@
+package com.braintreepayments.api;
+
+public class BrowserSwitchException extends Exception {
+
+    BrowserSwitchException(String message) {
+        super(message);
+    }
+}
