--- browser-switch/src/main/java/com/braintreepayments/browserswitch/BrowserSwitchException.java
@@ -1,8 +1,0 @@
-package com.braintreepayments.browserswitch;
-
-public class BrowserSwitchException extends Exception {
-
-    BrowserSwitchException(String message) {
-        super(message);
-    }
-}
