+++ browser-switch/src/test/java/com/braintreepayments/api/test/TestActivity.java
@@ -1,0 +1,6 @@
+package com.braintreepayments.api.test;
+
+import androidx.fragment.app.FragmentActivity;
+
+public class TestActivity extends FragmentActivity {
+}
