--- browser-switch/src/test/java/com/braintreepayments/browserswitch/test/TestActivity.java
@@ -1,6 +1,0 @@
-package com.braintreepayments.browserswitch.test;
-
-import androidx.fragment.app.FragmentActivity;
-
-public class TestActivity extends FragmentActivity {
-}
