+++ demo/src/main/java/com/braintreepayments/api/demo/DemoActivity.java
@@ -1,0 +1,24 @@
+package com.braintreepayments.api.demo;
+
+import android.os.Bundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+
+public class DemoActivity extends AppCompatActivity {
+
+    private static final String FRAGMENT_TAG = DemoFragment.class.getSimpleName();
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        FragmentManager fm = getSupportFragmentManager();
+        if (fm.findFragmentByTag(FRAGMENT_TAG) == null) {
+            fm.beginTransaction()
+                    .add(android.R.id.content, new DemoFragment(), FRAGMENT_TAG)
+                    .commit();
+        }
+    }
+}
