--- demo/src/main/java/com/braintreepayments/browserswitch/demo/DemoActivity.java
@@ -1,24 +1,0 @@
-package com.braintreepayments.browserswitch.demo;
-
-import android.os.Bundle;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.FragmentManager;
-
-public class DemoActivity extends AppCompatActivity {
-
-    private static final String FRAGMENT_TAG = DemoFragment.class.getSimpleName();
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        FragmentManager fm = getSupportFragmentManager();
-        if (fm.findFragmentByTag(FRAGMENT_TAG) == null) {
-            fm.beginTransaction()
-                    .add(android.R.id.content, new DemoFragment(), FRAGMENT_TAG)
-                    .commit();
-        }
-    }
-}
